Restructure the Nes project
This restructures the overall project layout to move away from being a
single win32 console app and towards being an all purpose nes library
available to different app platforms.

1. Seperate build files from core code. The win32 dll build files are in
src/nes/win32. These build files now build nes.dll which exports a C style
interface.

2. Create include/nes_api.h and expose core Nes API.

3. Seperate SDL code from nes core and remove IGfx and IHostInput
interfaces. The Sdl code should still be cross platform, so create a win32
specific build of it under src/sdl/win32. This console app is identical in
function to the old nes.exe.

TODO:
Create C++ interface for nes and export it.

+++ src/input.cpp
@@ -44,7 +44,7 @@ void Input::storeb(u16 addr, u8 val)
     }
 }
 
-IStandardController* Input::GetStandardController(u8 port)
+IStandardController* Input::GetStandardController(unsigned int port)
 {
     if (port >= 2)
     {
